+++ CHANGELOG.md
@@ -23,6 +23,7 @@ For older changes see the [archived Singularity change log](https://github.com/a
 
 - Set the `--net` flag if `--network` or `--network-args` is set rather
   than silently ignoring them if `--net` was not set.
+- Do not hang on pull from http(s) source that doesn't provide a content-length.
 
 ## v1.1.3 - \[2022-10-25\]